move enterprise tests into long
