+++ test/minikube/minikube_crash_handling_test.go
@@ -201,3 +201,36 @@ func TestReadWriteManyMultitenancy(t *testing.T) {
 		},
 	})
 }
+
+func TestNuoDBKubeDiagnostics(t *testing.T) {
+	testlib.AwaitTillerUp(t)
+	defer testlib.VerifyTeardown(t)
+	defer testlib.Teardown(testlib.TEARDOWN_ADMIN)
+
+	helmChartReleaseName, namespaceName := testlib.StartAdmin(t, &helm.Options{}, 1, "")
+
+	admin0 := fmt.Sprintf("%s-nuodb-cluster0-0", helmChartReleaseName)
+
+	defer testlib.Teardown(testlib.TEARDOWN_DATABASE) // ensure resources allocated in called functions are released when this function exits
+
+	databaseHelmChartReleaseName := testlib.StartDatabase(t, namespaceName, admin0, &helm.Options{
+		SetValues: map[string]string{
+			"database.sm.resources.requests.cpu":    testlib.MINIMAL_VIABLE_ENGINE_CPU,
+			"database.sm.resources.requests.memory": testlib.MINIMAL_VIABLE_ENGINE_MEMORY,
+			"database.te.resources.requests.cpu":    testlib.MINIMAL_VIABLE_ENGINE_CPU,
+			"database.te.resources.requests.memory": testlib.MINIMAL_VIABLE_ENGINE_MEMORY,
+		},
+	})
+
+	config := testlib.GetNuoDBK8sConfigDump(t, namespaceName, admin0)
+
+	assert.True(t, func() bool { _, ok := config.Pods[admin0]; return ok }())
+
+	tePodNameTemplate := fmt.Sprintf("te-%s-nuodb-%s-%s", databaseHelmChartReleaseName, "cluster0", "demo")
+	tePodName := testlib.GetPodName(t, namespaceName, tePodNameTemplate)
+	assert.True(t, func() bool { _, ok := config.Pods[tePodName]; return ok }())
+
+	smPodTemplate := fmt.Sprintf("sm-%s-nuodb-%s-%s", databaseHelmChartReleaseName, "cluster0", "demo")
+	smPodName := testlib.GetPodName(t, namespaceName, smPodTemplate)
+	assert.True(t, func() bool { _, ok := config.Pods[smPodName]; return ok }())
+}
